Convert shrinker2 to replace shrinker and update tests
